 requirements.txt
Flask>=2.0
pymongo>=4.0
requests>=2.28
apscheduler>=3.9
python-dotenv>=1.0
